968 - controller not updated
The problem with the logging/new controller not being updated
following a logging/:id route was caused by not passing
the controller into render. Because we were calling render with
the 'logging' template name, it was using the logging/:id route
controller which, of course, wasn't cleared upon arrival at the
logging/new route.

Solution is to pass it the controller as an option:
 this.render('logging', { controller: controller })

Interestingly, we were doing it right in the choice/new route.